added controller

+++ src/main/java/com/firstproject/learningms/repository/JpaRepo.java
@@ -3,6 +3,6 @@ package com.firstproject.learningms.repository;
 import com.firstproject.learningms.model.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaRepo extends JpaRepository<Books,Long> {
+public interface JpaRepo extends JpaRepository<Books,Integer> {
 
 }
